bugfix/Reliability/sonar
